+++ app/src/main/java/com/kiwi/kiwitalk/ui/keyword/recyclerview/SelectedKeywordAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.kiwi.domain.model.keyword.Keyword
+import com.kiwi.domain.model.Keyword
 import com.kiwi.kiwitalk.databinding.ItemKeywordBinding
 
 class SelectedKeywordAdapter() :
