+++ data/src/main/java/com/kiwi/data/repository/SearchChatRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.kiwi.data.datasource.remote.SearchChatRemoteDataSource
 import com.kiwi.domain.model.ChatInfo
 import com.kiwi.domain.model.Marker
 import com.kiwi.domain.model.PlaceChatInfo
-import com.kiwi.domain.model.keyword.Keyword
+import com.kiwi.domain.model.Keyword
 import com.kiwi.domain.repository.SearchChatRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
