+++ data/src/main/java/com/kiwi/data/repository/SearchKeywordRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.kiwi.data.repository
 
 import com.kiwi.data.datasource.remote.SearchKeywordRemoteDataSource
-import com.kiwi.domain.model.keyword.KeywordCategory
+import com.kiwi.domain.model.KeywordCategory
 import com.kiwi.domain.repository.SearchKeywordRepository
 import javax.inject.Inject
